+++ core/CHANGELOG.md
@@ -1,3 +1,25 @@
+# 1.1.0
+* Small comment/doc changes from weekly review
+* Easy fix to sort `api_container_env_vars` alphabetically
+* Remove some now-unneeded TODOs
+* Fix the `build_and_run.sh` script to use the proper way to pass in Docker args
+* Disallow test names with our test name delimiter: `,`
+* Create an access controller with basic license auth
+* Connect access controller to auth0 device authorization flow
+* Implement machine-to-machine authorization flow for CI jobs
+* Bind-mount Kurtosis home directory into the initializer image
+* Drop default parallelism to `2` so we don't overwhelm slow machines (and users with fast machines can always bump it up)
+* Don't run `validate` workflow on `develop` and `master` branches (because it should already be done before merging any PRs in)
+* Exit with error code of 1 when `build_and_run.sh` receives no args
+* Make `build_and_run.sh` also print the logfiles of the build threads it launches in parallel, so the user can follow along
+* Check token validity and expiration
+* Renamed all command-line flags to the initializer's `main.go` to be `UPPER_SNAKE_CASE` to be the same name as the corresponding environment variable passed in by Docker, which allows for a helptext that makes sense
+* Added `SHOW_HELP` flag to Kurtosis initializer
+* Switched default Kurtosis loglevel to `info`
+* Pull Docker logs directly from the container, removing the need for the `LOG_FILEPATH` variable for testsuites
+* Fixed bug where the initializer wouldn't attempt to pull a new token if the token were beyond the grace period
+* Switch to using `permissions` claim rather than `scope` now that RBAC is enabled
+
 # 1.0.3
 * Fix bug within CircleCI config file